requirements.txt
asgiref==3.11.1
Django==6.0.4
djangorestframework==3.17.1
django-cors-headers==4.5.0
openpyxl==3.1.2
reportlab==4.0.4
python-dotenv==1.1.0
requests==2.32.3
sqlparse==0.5.5
tzdata==2026.1
